+++ src/main/java/ezmarket/MemberMapper.java
@@ -53,21 +53,21 @@ public interface MemberMapper {
         List<MemberDTO> getAllUsersMember();
         
         //사용자 강퇴
-        void kickMember(long member_id, String member_kick_comment);
+        void kickMember(int member_id, String member_kick_comment);
         
         //사용자 복구
-        void restoreMember(long member_id, String member_kick_comment);
+        void restoreMember(int member_id, String member_kick_comment);
         
         //판매자 목록 가져오기
         List<MemberDTO> getAllBrandsMember();
         
         //판매자 승인
-        void sellApplicationAccept(long brand_id);
-        void sellApplicationAcceptAuthor(long brand_id);
+        void sellApplicationAccept(int brand_id);
+        void sellApplicationAcceptAuthor(int brand_id);
         
         //판매자 거절
-        void sellApplicationRefuse(long brand_id, String brand_refusal_comment);
-        void sellApplicationRefuseAuthor(long brand_id, String brand_refusal_comment);
+        void sellApplicationRefuse(int brand_id, String brand_refusal_comment);
+        void sellApplicationRefuseAuthor(int brand_id, String brand_refusal_comment);
         
         //판매자 주소 가져오기
 		ArrayList<MemberDTO> getBrandAddr();
